Limit to use uint16 range

+++ property-based-tests/numeral.go
@@ -3,7 +3,7 @@ package numeral
 import "strings"
 
 // ConvertToRoman converts an Arabic number to a Roman Numeral.
-func ConvertToRoman(arabic int) string {
+func ConvertToRoman(arabic uint16) string {
 	var result strings.Builder
 
 	for _, numeral := range allRomanNumerals {
@@ -17,8 +17,8 @@ func ConvertToRoman(arabic int) string {
 }
 
 // ConvertToArabic converts a Roman Numeral to an Arabic number.
-func ConvertToArabic(roman string) int {
-	var arabic int
+func ConvertToArabic(roman string) uint16 {
+	var arabic uint16
 
 	for _, numeral := range allRomanNumerals {
 		for strings.HasPrefix(roman, numeral.Symbol) {
@@ -31,7 +31,7 @@ func ConvertToArabic(roman string) int {
 }
 
 type romanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
